+++ PaintEditorWindow.c
@@ -688,7 +688,7 @@ void ww_setZoom(WINFO* wp, float newFactor) {
 /*
  * Increase / decrease the zoom factor of the current window. 
  */
-int ww_zoomWindow(int anIncreaseFactor) {
+long long ww_zoomWindow(int anIncreaseFactor) {
 	WINFO* wp = ww_getCurrentEditorWindow();
 
 	if (wp == 0) {
